separate the routes types declarations from the routes/*.ts code, to minimize bundling errors
