+++ server/package.json
@@ -22,9 +22,9 @@
     "prepack": "tsc --esModuleInterop genome/*.ts dataset/*.ts && webpack --env NODE_ENV=production",
     "response": "nodemon modules/test/test.server.js --watch src",
     "getconf": "../build/getConfigProp.js",
+    "checkers": "typia generate --input shared/types/routes/checkers/raw --output shared/types/routes/checkers/transformed --project ./shared/types/routes/checkers/tsconfig.json",
     "predoc": "npm run checkers",
-    "doc": "typedoc --json ../public/docs/server.json",
-    "checkers": "typia generate --input shared/checkers/raw --output shared/checkers/transformed --project ./shared/checkers/tsconfig.json"
+    "doc": "typedoc --json ../public/docs/server.json"
   },
   "author": "",
   "license": "SEE LICENSE IN ./LICENSE",
