+++ server/routes/genelookup.ts
@@ -45,14 +45,3 @@ export const api: any = {
 		}
 	}
 }
-
-export type GeneLookupRequest = {
-	input: string
-	genome: string
-	deep: boolean
-}
-
-export type GeneLookupResponse = {
-	error?: string
-	hits: string[]
-}
