+++ server/routes/healthcheck.ts
@@ -26,47 +26,3 @@ export const api = {
 		}
 	}
 }
-
-/**
- * for documentation only, to signify integer: not type-checked statically
- */
-type int = number
-
-/**
- * Information aboute the server build version and dates,
- * including the date when the server was last launched
- */
-export type VersionInfo = {
-	pkgver: string
-	codedate: string
-	launchdate: string
-}
-
-type BuildByGenome = {
-	[index: string]: GenomeBuildInfo
-}
-
-export type GenomeBuildInfo = {
-	genedb: DbInfo
-	termdbs?: {
-		[index: string]: DbInfo
-	}
-}
-
-type DbInfo = {
-	buildDate: string // "unknown" or a Date-convertible string
-	tables?: {
-		[index: string]: int
-	}
-}
-
-/**
- * Server sttus and data related to it's health
- */
-export type HealthCheckResponse = {
-	status: 'ok' | 'error'
-	genomes: BuildByGenome
-	versionInfo: VersionInfo
-	w?: number[]
-	rs?: number
-}
