--- server/shared/checkers/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": { 
-    "esModuleInterop": true,
-    "baseUrl": "./",
-    "strictNullChecks": true,
-    "skipLibCheck": true,
-    "noImplicitAny": false,
-    "plugins": [{
-      "transform": "typia/lib/transform"
-    }]
-  },
-  "exclude": [
-    "node_modules",
-    "**/*+(.spec|.e2e).ts",
-    "**/Point.ts"
-  ]
-}
