+++ server/shared/types/routes/genelookup.ts
@@ -1,0 +1,10 @@
+export type GeneLookupRequest = {
+	input: string
+	genome: string
+	deep: boolean
+}
+
+export type GeneLookupResponse = {
+	error?: string
+	hits: string[]
+}
